refactor(cmd): simplify flag handling & init logic
- Rename flags for consistency: `Lang` to `LangFlg`, `GitDir` to `GitDirFlg`, and `WorkTree` to `WorkTreeFlg`.
- Consolidate command registration in `root.go` by removing redundant `rootCmd.AddCommand` calls from individual command files (add.go, remove.go, list.go, doc.go).
- Simplify `init()` functions in command files to centralize logic.
- Update root command version handling to dynamically use `version` variable.
- Clarify documentation comment in `main.go`.
