+++ cmd/main.go
@@ -1,8 +1,6 @@
 /*
 Package cmd initializes all commands (including root command) for the HkUp CLI.
-
-Additionally, the package holds all tests for commands.
 */
 package cmd
 
-// Note: This file should be kept empty.
+// NOTE: This file is for documentation purposes and should be kept empty.
